compatibility-tests/sdk-test-python/requirements.txt
pytest==8.3.3
azure-storage-blob==12.23.1
azure-storage-queue==12.12.0
azure-data-tables==12.6.0